requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
python-multipart>=0.0.6
supabase>=2.0.0
